Remove columned chatbot. Refactor
