+++ frontends/main/src/page-components/LearningResourceDrawer/LearningResourceDrawer.tsx
@@ -1,6 +1,6 @@
 import React, { Suspense, useEffect, useId, useMemo } from "react"
 import { RoutedDrawer, imgConfigs } from "ol-components"
-import { LearningResourceExpanded } from "./LearningResourceExpanded"
+import { LearningResourceExpanded } from "../LearningResourceExpanded/LearningResourceExpanded"
 import type {
   LearningResourceCardProps,
   RoutedDrawerProps,
