+++ frontends/main/src/page-components/LearningResourceExpanded/AiChatSyllabus.test.tsx
