+++ frontends/main/src/page-components/LearningResourceExpanded/AiChatSyllabus.tsx
