+++ frontends/main/src/page-components/LearningResourceExpanded/testUtils.ts
